==> Day 1/01-sentiment-detection.ipynb <==
# Basic BERT models

BERT models can be used on their own to perform missing word prediction.
This notebook show in the first part how this works with the original
BERT model from Google. Abstractions have evolved, therefore we use
`pipeline` to accomplish that.

Apart from predicting missing words, BERT-like models are also suitable
for classification. The second part in this notebook uses a text classification
pipeline to perform (multilingual) sentiment detection.

## BERT models are Masked-Language models

They can guess missing words depending on the context they appear in!

from transformers import pipeline

unmasker = pipeline('fill-mask', model='bert-base-uncased')

unmasker("[MASK] is the capital of France.")

## Sentiment detection

As a simple example, BERT models can be used for sentiment detection.

Find more at https://huggingface.co/models?pipeline_tag=text-classification&sort=trending&search=sentiment


# Load the classification pipeline with the specified model
sentiment = pipeline("text-classification", 
                     model="nlptown/bert-base-multilingual-uncased-sentiment", 
                     top_k=None)

# Classify a new sentence
sentiment("I love O'Reilly online courses. I can learn a lot.")

sentiment("I love O'Reilly online courses! I can learn a lot!")

sentiment("I have to work so hard to learn this LLM stuff.")

sentiment("O'Reilly Kurse gefallen mir sehr gut. Ich kann dabei viel lernen!")
